Adding dist
