+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "toggle-device",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -14,7 +14,9 @@
     "@emotion/styled": "^11.11.5",
     "@mui/material": "^5.15.16",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@types/chrome": "^0.0.266",
