+++ src/index.css
@@ -13,3 +13,18 @@ body {
   height: auto;
   color: #163300;
 }
+
+ul {
+    text-align: left;
+}
+
+blockquote {
+    padding: 5px;
+    background-color: #9FE870;
+    border-radius: 25px;
+    margin-top: 100px;
+}
+
+a {
+    color: #fe2c96
+}
