+++ src/main.tsx
@@ -1,14 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom/client'
-import App from './App.tsx'
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { HashRouter as Router } from 'react-router-dom';
+
 import { ThemeProvider } from '@mui/material/styles';
-import './index.css'
+import App from './App.tsx';
+import './index.css';
 import theme from './theme.ts';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ThemeProvider theme={theme}>
-    <App />
+      <Router>
+        <App />
+      </Router>
     </ThemeProvider>
-  </React.StrictMode>,
-)
+  </React.StrictMode>
+);
