+++ vite.config.ts
@@ -1,8 +1,8 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
 
 // https://vitejs.dev/config/
 export default defineConfig({
+   base: '/',
   plugins: [react()],
-    base: process.env.NODE_ENV === 'production' ? '/device-toggle-chrome-ext/' : '/',
-})
+});
